requirements.txt
# Commodities Price Forecasting - Requirements
# State-of-the-art ML/DL dependencies (2024-2025)

# Core ML/DL
torch>=2.0.0
pytorch-lightning>=2.0.0
pytorch-forecasting>=1.0.0  # TFT implementation
neuralforecast>=1.7.0       # N-BEATS, N-HiTS (Nixtla)

# Traditional ML
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Data & Feature Engineering
pandas>=2.1.0
numpy>=1.24.0
ta>=0.11.0                  # Technical Analysis library

# Interpretability
shap>=0.44.0
captum>=0.6.0               # PyTorch model interpretability

# Experiment Tracking & Optimization
mlflow>=2.8.0
optuna>=3.4.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Utilities
tqdm>=4.66.0
joblib>=1.3.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Development
pytest>=7.4.0
black>=23.11.0
# NLP & Multimodal
transformers>=4.30.0
sentence-transformers>=2.2.0
newsapi-python>=0.2.7
openai>=1.0.0

